change the way for selecting recommendation posts I MODIFIED DATABASE so CHECK IT FIRST!!! Added reaction buttons (like and dislike) for each post and linked the likes with the output post. Now amount of likes and coefficient affects the placement of posts in recommendations. I recommend to you to DROP your schema by your hands before starting. P.S. Good luck sorting this shit out :)

+++ database/queries/delete_queries.py
@@ -1,15 +1,14 @@
 from sqlalchemy import select
 
 from config.logging_config import logger
-from database.models import Session, PersonalChannel, UserChannel, UserCategory, PersonalPost, GeneralChannel
+from database.models import Session, PersonalChannel, UserChannel, UserCategory, PersonalPost, PremiumChannel
 
 
 async def delete_personal_channel(username):
     session = Session()
     try:
-        personal_channel_id = \
-        session.execute(select(PersonalChannel.id).where(PersonalChannel.username == username)).fetchone()[0]
-        session.query(PersonalPost).filter(PersonalPost.channel_id == personal_channel_id).delete()
+        personal_channel_id = session.execute(select(PersonalChannel.id).where(PersonalChannel.username == username)).fetchone()[0]
+        session.query(PersonalPost).filter(PersonalPost.personal_channel_id == personal_channel_id).delete()
         session.flush()
         session.query(UserChannel).filter(UserChannel.channel_id == personal_channel_id).delete()
         session.flush()
@@ -26,8 +25,7 @@ async def delete_personal_channel(username):
 async def delete_user_category(user_tg_id, category_id):
     session = Session()
     try:
-        session.query(UserCategory).filter(UserCategory.user_id == user_tg_id,
-                                           UserCategory.category_id == category_id).delete()
+        session.query(UserCategory).filter(UserCategory.user_id == user_tg_id, UserCategory.category_id == category_id).delete()
         session.commit()
         return True
     except Exception as err:
@@ -40,7 +38,7 @@ async def delete_user_category(user_tg_id, category_id):
 async def delete_general_channel(channel_username):
     session = Session()
     try:
-        session.query(GeneralChannel).filter(GeneralChannel.username == channel_username).delete()
+        session.query(PremiumChannel).filter(PremiumChannel.username == channel_username).delete()
         session.commit()
         return True
     except Exception as err:
